refactor: refactor AuthPage component and auth forms
